requirements.txt
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.9
pytesseract==0.3.9
